Delegate transaction refactor
* use cacheTransactionSummaries to avoid race conditions
* simplify submit handler in ConfirmTransactionDialog
* remove unnecessary calculatingDelegationFee checks
* move 'deregisterStakingKey' into delegate file
